+++ src/main/java/com/jasik/momsnaggingapi/domain/question/controller/QuestionController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.question.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class QuestionController {
+}
