+++ src/main/java/com/jasik/momsnaggingapi/domain/question/service/QuestionService.java
@@ -1,0 +1,13 @@
+package com.jasik.momsnaggingapi.domain.question.service;
+
+import com.jasik.momsnaggingapi.domain.question.repository.QuestionRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class QuestionService {
+    private final QuestionRepository questionRepository;
+}
